fix(clustering): 修模組驗證的靜態分析與圈複雜度，補候選契約測試
模組驗證（scripts/verify_module.py）三項未達標，逐項處理：

**靜態分析**（新增行 5 個 → 0）：RUF046 多餘的 int(round())、I001 import 排序、
RUF100 未啟用的 noqa、S110 try-except-pass（改為記錄 warning——coherence 算不
出來不影響其餘三項指標，但要留痕跡，否則「為什麼這次沒有一致性分數」查不出來）。

**圈複雜度**（我新寫的 3 支超 B → 全部降到 B 以下）：
- engine._evaluate：判準改成資料表 _CRITERIA，加判準時只需多一列
- keywords.extract_top_terms：拆出詞頻統計、DF 計算、單群計分
- model.rank_candidates：拆出單一指標欄的正規化

**結構修正**：掃描腳本原本自己定義了一份判準門檻，那是同一份知識的第二個落點
——改成一律 import engine 的常數。⚠ 這個腳本是拿來驗證正式流程的，門檻若兩處
各寫一份，驗的就不是正式流程了。

**候選契約**（實機驗收抓到）：DP-Means 候選缺 candidate_k（finalize 與 API 都讀
它，我憑印象寫成 k），且 k_scan 落在 state 頂層而非 metrics.k_scan。兩者都是
「不會報錯、只是東西讀不到」的類型，補 test_dpmeans_candidate_contract.py 釘住。

**.env 載入**：驗收腳本在連線前沒載入 .env，拿到預設的 localhost:5433（兩天前
就停掉的容器）。⚠ 這個載入原本藏在 runner.py 的 import side effect 裡——任何
「連線發生在 import runner 之前」的腳本都會靜默連到錯的地方，症狀是連線逾時，
看起來像 DB 掛了。

+++ backend/app/clustering/dpmeans.py
@@ -142,7 +142,7 @@ def derive_lambda(sample: list[Vector]) -> LambdaResult:
         for j in range(i + 1, len(points))
     )
     index = min(len(distances) - 1,
-                max(0, int(round(PAIRWISE_QUANTILE * (len(distances) - 1)))))
+                max(0, round(PAIRWISE_QUANTILE * (len(distances) - 1))))
     value = distances[index]
     if value <= 0.0:
         # 全部重複文件時距離會是 0——那不是有效門檻。
